pyproject.toml
[project]
name = "agentic-ai-copilot"
version = "0.1.0"
description = "An AI copilot project"
dependencies = [
    "langchain-core",
    "langchain-community",
    "langchain-openai",
    "openai",
    "pydantic",
    "requests",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "crewai[tools]",
    "crewai",
]


[tool.uv]
